inverse matrix

+++ backend/matrix3x3.h
@@ -2,7 +2,7 @@
 
 struct Matrix3x3
 {
-	float values_[3][3];
+	double values_[3][3];
 
 	bool inverse(Matrix3x3& result);
 };
